@@ -2,23 +2,19 @@
   "name": "client",
   "version": "0.0.1",
   "devDependencies": {
-    "@cartamd/plugin-code": "^4.0.5",
-    "@cartamd/plugin-emoji": "^4.0.2",
     "@sveltejs/adapter-auto": "^3.0.0",
-    "@sveltejs/kit": "^2.0.0",
-    "@sveltejs/vite-plugin-svelte": "^3.0.0",
+    "@sveltejs/kit": "^2.5.27",
+    "@sveltejs/vite-plugin-svelte": "^4.0.0",
     "autoprefixer": "^10.4.19",
-    "carta-md": "^4.2.0",
     "daisyui": "latest",
     "isomorphic-dompurify": "^2.12.0",
-    "mdsvex": "^0.11.2",
     "postcss": "^8.4.38",
-    "svelte": "^4.2.7",
-    "svelte-check": "^3.6.0",
+    "svelte": "^5.0.0",
+    "svelte-check": "^4.0.0",
     "tailwindcss": "^3.4.4",
     "tslib": "^2.4.1",
-    "typescript": "^5.0.0",
-    "vite": "^5.0.3"
+    "typescript": "^5.5.0",
+    "vite": "^5.4.4"
   },
   "private": true,
   "scripts": {
@@ -30,9 +26,12 @@
   },
   "type": "module",
   "dependencies": {
+    "@shikijs/rehype": "^1.26.2",
     "caniuse-lite": "^1.0.30001690",
     "prettier": "^3.3.2",
-    "prettier-plugin-svelte": "^3.2.5",
-    "prettier-plugin-tailwindcss": "^0.6.5"
+    "prettier-plugin-svelte": "^3.2.6",
+    "prettier-plugin-tailwindcss": "^0.6.5",
+    "shiki": "^1.27.0",
+    "svelte-exmarkdown": "^4.0.2"
   }
 }
