@@ -1,5 +1,5 @@
 <script lang="ts">
-    export let data;
+    let { data } = $props();
 </script>
 
 <p>{data.info}</p>
