@@ -2,11 +2,14 @@
     import "../../app.css";
     import type { ActionData } from "./$types";
 
-    export let form: ActionData;
+    interface Props {
+        form: ActionData;
+    }
+
+    let { form }: Props = $props();
 </script>
 
-<div
-    class="gird-cols-1 bg-95 grid min-h-screen w-full items-center justify-center gap-2">
+<div class="max-h-95 bg-95 grid min-h-95 w-full items-center justify-center overflow-scroll">
     <form class="grid bg-white" action="/register" method="POST">
         <p class="m-3 text-xl font-bold">Register</p>
         <input
@@ -29,6 +32,11 @@
             class="txt-in95 m-3"
             name="password"
             placeholder="Password" />
+        <input
+            type="password"
+            class="txt-in95 m-3"
+            name="verifyPassword"
+            placeholder="Verify Password" />
         <button type="submit" class="btn95 m-3 max-w-min"
             ><div>Submit</div></button>
         {#if form?.failure}
