@@ -2,9 +2,13 @@
     import "../app.css";
 </script>
 
-<a href="login">Login</a>
-<a href="register">Register</a>
-<a href="info">Info</a>
-<a href="blog">Blog</a>
+<div class="bg-white min-w-full p-2">
+<a class="btn95 m-2" href="/">Home</a>
+<a class="btn95 m-2" href="/login">Login</a>
+<a class="btn95 m-2" href="/register">Register</a>
+<a class="btn95 m-2" href="/info">Info</a>
+<a class="btn95 m-2" href="/blog/post">Make Post</a>
+<a class="btn95 m-2" href="/blog">Blog</a>
+</div>
 
 <slot />
